// src/menu-entities/ingredients/types/measurement.enum.ts
/**
 * Measurement Enum
 *
 * Defines standard measurement units for ingredients
 */
enum Measurement {
  KILOGRAM = 'kg',
  GRAM = 'g',
  LITER = 'l',
  MILLILITER = 'ml',
  PIECE = 'pc',
  TEASPOON = 'tsp',
  TABLESPOON = 'tbsp',
  CUP = 'cup',
  OUNCE = 'oz',
  POUND = 'lb',
  GALLON = 'gal',
  QUART = 'qt',
  PINT = 'pt',
}

export default Measurement;